Switch from IntegerField to BigIntegerField to better accomodate potential large amounts

+++ platalbank_core/models/account.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class Account(models.Model):
-    balance = models.IntegerField() # Number of cents owned
+    balance = models.BigIntegerField() # Number of cents owned
     description = models.CharField(max_length=1024)
     short_name = models.CharField(max_length=128)
     # TODO: owner = models.ForeignKey("LegalPerson")
